+++ runtime/common/Cargo.toml
@@ -1,0 +1,94 @@
+[package]
+name = "runtime-common"
+authors = { workspace = true }
+description = "Common implementations for all tanssi runtimes"
+edition = "2021"
+license = "GPL-3.0-only"
+version = "0.1.0"
+
+[package.metadata.docs.rs]
+targets = [ "x86_64-unknown-linux-gnu" ]
+
+[dependencies]
+hex-literal = { workspace = true }
+parity-scale-codec = { workspace = true, features = [ "derive" ] }
+scale-info = { workspace = true, features = [ "derive" ] }
+
+# Own
+dp-core = { workspace = true }
+pallet-configuration = { workspace = true }
+pallet-data-preservers = { workspace = true }
+pallet-invulnerables = { workspace = true }
+pallet-pooled-staking = { workspace = true }
+pallet-registrar = { workspace = true }
+pallet-services-payment = { workspace = true }
+
+# Moonkit
+pallet-migrations = { workspace = true }
+
+# Substrate
+frame-try-runtime = { workspace = true, optional = true }
+frame-support = { workspace = true }
+frame-system = { workspace = true }
+pallet-balances = { workspace = true }
+
+sp-core = { workspace = true }
+sp-runtime = { workspace = true }
+sp-std = { workspace = true }
+
+# Cumulus
+cumulus-primitives-core = { workspace = true }
+
+[build-dependencies]
+substrate-wasm-builder = { workspace = true }
+
+[features]
+default = [
+	"std",
+]
+std = [
+	"cumulus-primitives-core/std",
+	"frame-support/std",
+	"frame-system/std",
+	"pallet-balances/std",
+	"pallet-configuration/std",
+	"pallet-data-preservers/std",
+    "pallet-invulnerables/std",
+	"pallet-migrations/std",
+	"pallet-registrar/std",
+    "pallet-pooled-staking/std",
+	"pallet-services-payment/std",
+	"parity-scale-codec/std",
+	"scale-info/std",
+	"sp-core/std",
+	"sp-std/std",
+]
+
+runtime-benchmarks = [
+	"frame-support/runtime-benchmarks",
+	"frame-system/runtime-benchmarks",
+	"pallet-balances/runtime-benchmarks",
+	"pallet-configuration/runtime-benchmarks",
+	"pallet-data-preservers/runtime-benchmarks",
+    "pallet-invulnerables/runtime-benchmarks",
+	"pallet-migrations/runtime-benchmarks",
+    "pallet-pooled-staking/runtime-benchmarks",
+	"pallet-registrar/runtime-benchmarks",
+	"pallet-services-payment/runtime-benchmarks",
+]
+
+try-runtime = [
+	"frame-support/try-runtime",
+	"frame-system/try-runtime",
+	"frame-try-runtime/try-runtime",
+	"pallet-balances/try-runtime",
+	"pallet-configuration/try-runtime",
+	"pallet-data-preservers/try-runtime",
+    "pallet-invulnerables/try-runtime",
+	"pallet-migrations/try-runtime",
+    "pallet-pooled-staking/try-runtime",
+	"pallet-registrar/try-runtime",
+	"pallet-services-payment/try-runtime",
+]
+
+fast-runtime = []
